+++ src/main/java/ladder/domain/Players.java
@@ -1,6 +1,5 @@
-package ladder;
+package ladder.domain;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
